Fixes CRUD for comments. Comments now add, delete and edit correctly. Comments section only appears for logged in users, Comment Edit and Delete buttons only appear for admin users. Controller has been locked down so that routes can only be used by authenticated users. Also images are renamed to not have the leading zeros, they now match their pokemon id's from the database. All pokemon.show pages load with a pokemon image.

+++ resources/views/pokemon/show.blade.php
@@ -2,39 +2,43 @@
 
 @section('content')
 
-	<!-- Main jumbotron for a primary marketing message or call to action -->
-	<div class="jumbotron">
-	  <div class="container">
-		<h1>Pokemon</h1>
-		<h2>{{ $pokemon->name }}</h2>
-		<img src="..\images\pokemon\00{{ $pokemon->id }}.png">
-		@foreach ($comments as $comment)
-			<p>
-				{{ $comment->comment }} - Posted by: {{ $comment->user->name }} 
-				<a href="{{ route('comments.edit', $comment->id) }}" class="btn btn-warning">Edit Comment</a>
-
-				{!! Form::open(['action' => ['CommentsController@destroy', $comment->id], 'method' => 'delete']) !!}
-					{!! Form::submit('Delete', ['class'=>'btn btn-danger']) !!}
-				{!! Form::close() !!}
-			</p>
-		@endforeach
-	  </div>
-	</div>
-
-	<div class="container">
-		<h4>Comment on {{ $pokemon->name }}</h4>
-		
-		{!! Form::open(['route' => ['comments.store', $pokemon->id], 'class' => 'form-horizontal', 'method' => 'POST']) !!}
-			<div class="form-group">
-				
-				{!! Form::label('Comment:') !!}
-				{!! Form::textarea('comment', $newComment->comment, ['class' => 'form-control']) !!}
+		<!-- Main jumbotron for a primary marketing message or call to action -->
+		<div class="jumbotron">
+			<div class="container">
+				<h1>Pokemon</h1>
+				<h2>{{ $pokemon->name }}</h2>
+				<img src="..\images\pokemon\{{ $pokemon->id }}.png">
+				@foreach ($comments as $comment)
+					<p>
+						{{ $comment->comment }} - Posted by: {{ $comment->user->name }}
+						
+						@if(Auth::check() && Auth::user()->role === "admin")
+							{!! Form::open(['action' => ['CommentsController@destroy', $comment->id], 'method' => 'delete']) !!}
+								{!! Form::submit('Delete', ['class'=>'btn btn-danger']) !!}
+								<a href="{{ route('comments.edit', $comment->id, $pokemon->id) }}" class="btn btn-warning">Edit Comment</a>
+							{!! Form::close() !!}
+						@endif
+					</p>
+				@endforeach
 			</div>
+		</div>
 
-			<div class="form-group">
-				{!! Form::submit('Add Comment', ['class' => 'btn btn-primary form-control']) !!}
-			</div>
-		{!! Form::close() !!}
-	</div>
+	@if(Auth::check())
+		<div class="container">
+			<h4>Comment on {{ $pokemon->name }}</h4>
+			
+			{!! Form::open(['route' => ['comments.store', $pokemon->id], 'class' => 'form-horizontal', 'method' => 'POST']) !!}
+				<div class="form-group">
+					
+					{!! Form::label('Comment:') !!}
+					{!! Form::textarea('comment', $newComment->comment, ['class' => 'form-control']) !!}
+				</div>
 
+				<div class="form-group">
+					{!! Form::submit('Add Comment', ['class' => 'btn btn-primary form-control']) !!}
+				</div>
+			{!! Form::close() !!}
+		</div>
+	@endif
+	
 @endsection
